PsychVulkanCore: Initial support for MS-Windows.
This adds the initial support code for MS-Windows Vulkan WSI.
On Windows we open our own win32 onscreen window for Vulkan
presentation, using its own Window class. Then we attach a win32
Vulkan surface to it, and use the fullscreen_exclusive extension to
switch to fullscreen mode when appropriate. It is important to not
even use the fullscreenExclusiveInfo struct in windowed mode, or
Vulkan will screw up big time.

This commit is the result of a nighmarish debug session, the amount
of fragility and sometimes absurd behaviour of Vulkan on Windows,
and  especially when interacting with OpenGL contexts on the same
thread is just mind-blowing, but not in a good way at all. E.g., part
of the Vulkan driver init must happen before creating a PTB onscreen
Window and its OpenGL context, if driver init happens after OpenGL
context creation, then that will damage the OpenGL context on AMD
drivers in a way that causes mind-boggling crashes! Go figure... Such
crashes don't happen on NVidia.

So this sort of works most of the time on AMD gpu's with these caveats:

- One out of five sessions the OpenGL  rendering will fail, with lots of
invalid operation GL errors. Restarting the script makes it work for a while.
Some  race-condition somewhere?!?

- On Matlab, the AMD driver can switch to fullscreen exclusive mode, but
on Octave it consistently fails to do so?!?

- On Octave, both installed gpu's are enumerated as Vulkan physical devices,
but on Matlab, only one is enumerated - the one also chosen for the OpenGL
rendering. No clue why?!?

- Trying to display with AMD's Vulkan driver to a NVidia connected display
crashes the AMD driver, despite the driver claiming it can do that. No such
behaviour for NVidia.

This commit only makes AMD work though. NVidia currently just renders a
black image. The Vulkan side of things works and can display basic test
patterns (flags = 1 test setting), but something goes wrong with OpenGL
interop. This needs more investigation.