PsychVulkanCore: Add missing setupfor interop image export.
The spec mandates having a VkExternalMemoryImageCreateInfo struct
chained to the VkImageCreateInfo for images whose backing memory
may be exported to other api's/driver instances/processes, as is the case
for OpenGL->Vulkan interop. So do it.

In practice, none of the  AMD or NVidia drivers on Linux or Windows
cared about us lacking this bit.