PsychVulkanCore: Use blit to flip OpenGL image into upright orientation.
Always use vkCmdBlitImage() for the interop->swapchain transfer, as this
allows us with some manipulation of the dstOffsets region to flip the image
vertically from OpenGL orientation into Vulkan orientation, so that it displays
upright. This allows us to avoid a vertical flip pass in PTB's imaging pipeline.

Even if a vkCmdBlitImage blit would be a bit less efficient than a pure copy
via vkCmdCopyImage, saving a complete buffer copy on the OpenGL side
should be a substantial saving in processing time.