PsychVulkanCore: Add DXGI fallback for HDR display queries on Windows.
The NVidia Vulkan driver on MS-Windows doesn't support queries of
HDR display properties in the nice way that the AMD drivers on Linux
and Windows do. Add a fallback implementation to deal with this
limitation.

We use the low-level DXGI 1.6 api to find the  DXGI output which
corresponds to the Vulkan display window and query that output
for the HDR display properties. This works fine as tested on Windows
with AMD and NVidia. This is a fallback, so currently only gets used
for NVidia gpu's.

As the AMD Vulkan driver has a bug in its HDR display query code,
we may need to enable this also unconditionally for AMD gpu's if
new AMD driver releases don't fix the bug in time. The bug has been
reported to the Windows team, but i don't have any feedback if it
has been fixed, so time will tell.

-> As we need DXGI 1.6 for this fallback, this means that
PsychVulkanCore will only work on Windows-10 version
1703 (10.0; Build 15063), also known as Windows 10
Creators Update.