PsychVulkan: Fixup basic diagnostic mode for non-interop drivers.
Last commit was flawed, this fixes it. Also change
PsychVulkanCore to output black-white alternating
images for double-buffering, or black-white-red-yellow
on triple-/quad-buffering. We never want to see anything
but black-white or timing at least without Vulkan timing
extensions will be broken. Black-White for double-buffering
is optimal to drive photo-diodes for basic timing tests.