Screen/Linux: Avoid redundant RandR output props query in non-VRR mode.
Avoid needless call to PsychOSGetOutputProps() in OpenWindow path
if vrr is not requested. This prevents an 'invalid outputIndex no
such output' abort on Vulkan multi-window configurations.

The inability to find RandR outputs that are logically disconnected
if a RandR output is leased out to Vulkan remains a big pita. We
can work around it to some degree with hacks, but i think i soon
will have to redesign RandR output enumeration in Screen(), as this
hackery has serious limits. Atm. we can't call any function that tries
to find a RandR output based on outputIndex, because we find RandR
outputs by mapping output indices to RandR crtc's and then find the
matching RandR output for a RandR crtc, an approach that fails badly
if no RandR crtc is associated with a logically disabled RandR output,
because the RandR output has been leased out to a direct display client
like our Vulkan driver.

Anyway, this hack makes the following commits (barely) work ok'ish,
good enough to get some testing done.