PsychVulkan: Enable basic ability to open multiple Vulkan windows.
On Linux, once a window is opened in fullscreen direct display mode,
the corresponding leased out RandR output becomes unavailable to
Screen(), so trying to access it via Screen('ConfigureDisplay','Scanout')
ends badly.

Therefore we keep track which outputs have been taken over in
direct display mode or fullscreen exclusive mode on Linux and
Windows, so we can avoid accessing them for enumeration/probing/
output selection/Window to output mapping etc. Together with a
small hack to Screen() for Linux, this allows to open multiple
PTB onscreen windows in Vulkan mode for theoretical multi-display
stimulation.

Ofc., this wouldn't be Vulkan if the drivers wouldn't be full
of bugs which make this more painful than what would expect:

- Mesa radv, at least as of Mesa 19.2 does not allow direct-
  display mode on more than one output at a time. On the
  2nd display it fails, and then is also dead on the first
  display it already took over! Yay! Now current radv does
  lack internal code to deal with this, i'd just would have
  hoped for a slightly more graceful failure mode :(

- amdvlk can deal with multiple display outputs, but when
  closing a dual-display setup, the 1st opened direct mode
  display will get stuck and can only be unstuck by opening
  and then closing another Vulkan window on the stuck output!
  How depressing is this?

Luckily multi-display operation is not in scope for the current
stage of the project, so i can defer proper solutions to a later
point in time. I just hoped for a quick win here, but nope...