PsychVulkan: Add workaround for radv on Mesa < 20.1.2 bug.
radv doesn't release its direct display mode window without
Vulkan instance destruction, aka full shutdown of PsychVulkanCore(),
neither does any other Mesa based driver, due to a bug that
i fixed. The bug fix is in Mesa 20.1.2 and later, but if i
want to allow for at least some basic testing on older Mesa
versions then i need a workaround. So we now detect if we
run on a Mesa based Vulkan driver in fullscreen mode on a
Mesa version older 20.1.2 and therefore buggy. If so, we
output a warning to the user to upgrade to 20.1.2+ and enable
the workaround which will close *all* Vulkan window via a full
driver shutdown if even one fullscreen window is open.

I guess for the final release build it will make sense to
disable use of radv by default, given the remaining bugs,
at least if an alternative AMD Vulkan driver is installed.