PsychVulkan: Adapt ScreenToHead mappings for direct-display-mode configs.
If usercode wants to use a single-X-Screen, multi-display setup with
multiple outputs per X-Screen and open a window on one output, then
our standard assumption that the primary (rank 0) output is the right
one as beamposition/timing reference for windows on a given X-Screen
no longer holds.

Adapt the mapping accordingly: As we know on which video output i our
Vulkan window will display fullscreen/direct-display, we can change
the ScreenToHead mapping dynamically so that output i becomes the
rank 0 output which Screen() uses for calibrations and as a timing
reference and for beamposition queries.

This way Screen will detect the correct VBLStartline, VBLEndline,
perform correct beamposition refresh calibration, use proper nominal
refresh interval queried from right RandR output, and use the correct
display engine for beamposition queries.

TODO: This is a bit of a hack and long term it would be better to
allow assigning a specific output index to the onscreen window and
then let Screen() use that index internally instead of all these
temporary remapping stuff. Could save a couple dozen microseconds
of overhead and be more robust in case of error-aborts.

Anyway, this seems to work well as tested on a single-X-Screen,
dual-display setup with Vulkan windows on each of the outputs,
letting all calibrations and tests pass and proper beamposition
queried.

In practice, proper mapping during startup is always useful to
let Screen() startup tests pass and calibrate properly. Proper
mapping during timestamping is only useful if beamposition
timestamping is used, which is only the case on NVidia with
binary blob driver, and as a rarely needed fallback for AMD
gpu's with DCE display engines. The only fallback case would
be a mirrored/cloned display setup, as in all other cases our
OS builtin vblank timestamping does a reliable and accurate job.