PsychVulkan: Load perfect identity gamma lut's in HDR mode.
For HDR display, we want the gamma lut's out of the way
to not mess with the HDR signal, so we need to load
identity lut's before opening the Vulkan window. We
can't use LoadIdentityClut(), as it uploads 8 bpc
truncated lut's in some configurations.

Therefore implement custom high precision y = x; identity
lut upload in PsychVulkan for the HDR case.