Screen: Fix colorimetry debug code for Screen('OpenMovie').
Make it query the 'src' pad of the actual video codec, so
we get caps with intact original colorimetry info, instead
of the caps already adapted (and devoid of meaningful
colorimetry info) to our appsink.

Fall back to old appsink 'sink' caps if the videocodec
element could not be found in the GStreamer pipeline.

-> Gives much better results on GStreamer 1.16, test on
GStreamer 1.17 pending.