HDR janitorial work: Move demos and functions into appropriate places.
The BrightSideDisplay support is as dead and obsolete as are likely
all remaining such display devices and the former company. After over
a decade, it is time to retire this stuff.

This commit moves our new HDR demos to the PsychDemos folder, and the
HDRRead() function to the PsychFiles folder. At also deletes some
obsolete files in the BrightSideDisplay folder.

This prepares the eventual removal of BrightSideDisplay support.