tinyexr.h: Add compile fix for Octave 5.2 / MinGW64 for MS-Windows.
For god knows what absurd reason, defining the variable
std::stringstream ss; defined in one if() scope of the subfunction
LoadDeepEXR() leads to linker failure with an undefined reference
to  std::__cxx11::basic_stringstream, despite the fact that the same
definition/reference is used in many similar places in the file.
Maybe some compiler/linker bug. Anyway, implement an alternative
solution to what that problematic code segment did.

Mixing C++ code with our C mex files just continues to be a frickin
pain in the ass...