PsychVulkanCore('HDRMetadata'): Skip redundant HDR metadata updates.
Check if the to-be-set metadata has really changed, ie. it isn't a redundant
update to the settings that were already current. Turns out this is important
on MS-Windows with both AMD and even more so NVidia drivers, or
performance will fall off a cliff! E.g., AMD takes over 100 msecs for an update,
and NVidia takes on the order of 1 second and a modesetting sequence with
complete signal dropout during that time! Strong work :/