ColorCal2: Make more robust against weirdly behaving display panels.
E.g., my Samsung C27HG70 will do something to its output light which
confuses the hell out of the CRS ColorCal2, as soon as the monitor
reaches or exceeds its rated peak luminance of 603 nits. Maybe some
kind of high-frequency flicker of the backlight? Visually all looks
the same to me, but the ColorCal's status led suggests lots of light
flashes like with a high frequency pulsing of the backlight?

This retries every measurement 3 times before giving up and marking
output as NaN invalid and s.reliable = 0. On USB based systems like
Linux and macOS it also tries a different type of parsing to try to
recover. This works, but gives bogus (way too low) readings, which
however generally follow the shape of the light pattern in HDR tests,
as if the whole luminance curve would be systematically offset down
by a few hundred nits.

MeasXYZ now also maps the new s.reliable = 0 flag to proper 'qual'
status codes 0 for success or -1 for trouble.