HDRTest: Some cleanup. Add test for primaries, whitepoint, gamut.
Three new tests test with a pure red, pure green and pure blue
patch to measure the location of the red, green and blue primaries
of the display. Also test 1 measures the white-point.

Then a nice plot is created which shows the color gamut and the
measured gamut triangle vs. display reported gamut triangle.