HDRTest: Also report 12 bpc quantized PQ values in addition to 10 bpc.
After endless hours of tweaking and rerunning the test, i finally
have a result that pleases me:

A mostly linear monitor response where measured nits for a pure
achromatic (white/grayscale) test patch of test 1 corresponds to
the actually defined and requested nits closely, with an error
of typically < 1.5% over most of the tested range up to 591 nits,
ie. 98% of rated peak luminance of the Samsunc CH27HG70 HDR monitor.

Still observe weird cutoff/crazyness of the monitor beyond the 98%
point instead of graceful rolloff, but this is workable.

Tested configuration that works:

- Ubuntu Linux 20.04.1 LTS, Linux kernel 5.8-rc7.

- AMD Raven Ridge gpu (Vega11 + DCN-1 display engine) on a
  Ryzen 5 2400G processor with integrated graphics.

- HDMI connection, with resolution/refresh rate reduced to
  1920x1080 at 60 Hz instead of native 2560x1440 at 144 Hz,
  because otherwise the HDMI bandwidth is insufficient for
  10 bpc signal transmission.

- All Freesync support completely disabled in the monitor OSD.

- Black equalizer to 20 (maximum). Picture mode "Custom".

- Brightness 100, Contrast 75, Sharpness 48

- Color Red,Green,Blue = 35,35,35 for color tone "Custom".

- Gamma "Mode1"

- Eye Saver "off"

- Low input lag "On" Panel refresh reported 144 Hz, but with
  video input from computer 60Hz.

- Local dimming "auto" ie. on in HDR mode.

-> This is the first setup that works reasonable, other configs
   may be possible. E.g., have to reverify if Displayport works,
   or FreeSync can be reenabled, or how Windows-10 fares.

   The key that made things much better was to set the color gains
   to 35,35,35 instead of their 50,50,50 factory default for "Normal"
   color tone. And maybe switching from Gamma mode 3 to 1.