PsychHDR: Allow passing new HDRMetadata also as a struct.
In addition to the separate parameters format, directly handled
by PsychVulkan() / PsychVulkanCore(), also allow passing in static
HDR metadata parameters as a struct 'newHdrMetadata' as in:

oldHdrMetadata = PsychHDR('HDRMetadata', window [, newHdrMetadata]);

This allows to pass modified versions of 'oldHdrMetadata' or of
metadata provided as struct by Screen('OpenMovie') directly, without
need to break it up into separate parameters. Iow. this is a
convenience function.

When we are at it, improve documentation of 'HDRMetadata', specify
valid range in nits and reference the standard that specifies the
content and meaning of HDR static metadata type 1.