PsychVulkanCore/PsychHDR: Make metadataType argument mandatory for setting.
In PsychVulkanCore('HDRMetadata', window, metadataType, ...) the
metadataType was so far optional and defaulted to the only currently
supported value 0 (Static HDR metadata Type 1). It seems to be better
for documentation of code and clarity to require explicit specification
of the value, at least for set calls. So now we require the parameter
on actual HDRMetadata set calls, ie. if more than 2 input args are
given, and only allow it to be optional on pure get calls, in which
case the returned structs oldHdrMetadata.MetadataType field will
specify the format of the returned data.

Update PsychHDR accordingly. Also some help text updates wrt. parameters.