Screen: Add support for decoding all known EOTF's as of GStreamer 1.18.
Add a switch-case statement to recognize and handle all movie EOTF's
currently recognized and suppported as of GStreamer 1.18.0, at this
time these are 16 known EOTF's, plus the "unknown" EOTF and default.

Add proper implementations, using the C code in GStreamer 1.18.0
as reference implementation, translated into GLSL. See:

https://gitlab.freedesktop.org/gstreamer/gst-plugins-base/-/blob/master/gst-libs/gst/video/video-color.c

For HDR decoding we now implement PQ and HLG, all other EOTF's are
for SDR movie decoding. Most of these EOTF's are not verified though,
difficult to do without suitable reference video footage.

PQ encoded content looks correct though, which is the most important
case for our purpose.

Note: The GLSL switch() statement was introduced in GLSL-1.3, which
strictly requires OpenGL 3.0 support. Apples trainwreck does not
really support OpenGL 3.0 in compatibility profile mode, which we
need. It may be though that they do support GLSL 1.3+ despite not
supporting OpenGL 3.0. Needs to be verified. Otherwise this would
be unsupported on macOS, or we'd need to rewrite the switch()
statement into many if() statements - a mild nightmare.