Screen: Use GStreamer range extension for format 11 instead of our own.
Turns out that after a bit of math, GStreamers limited-range/full-range
to normalized range mapping is mathematically identical to my own
solution, but a bit more general (allows separate bit depths per color
component) and elegant in terms of math. Switch to GStreamer method, so
we can outsource this and simplify our code and maintenance.