Screen: Fixup HDR PQ EOTF decode / make it more robust.
Checking our PQ EOTF decode shader - derived from GStreamer reference
C code and other sources - against the official BT 2100 spec shows we
(and GStreamer?) missed a max() operator for clamping out-of-range
values from turning into mayhem for the pow() function and potentially
creating NaN's or Inf(inities)! This was found when trying to track
down the visual artifacts in some HDR movies, not clear if the general
out-of-range stuff caused it, or this, but better safe (and spec
compliant!) than sorry!

-> Now our PQ EOTF decode is validated against the official mathematical
   definition.

-> Various other EOTF's were also checked wrt. their implementation
   and math against available specs, or other public sources like tutorials
   or Wikipedia, when available. Not all EOTF's are checked though,
   because missing info. Many are not tested, because missing reference
   movie footage. But this is as good as it gets for the initial release.

=> Here goes another almost 20 work hours for bug hunting in specs
   and reference implementations and what not!