Add helper functions for XYZ->RGB and RGB->XYZ colorspace conversion.
XYZToRGBMatrix() and RGBToXYZMatrix() are our newbies.

These can map from a specified RGB colorspace to XYZ space, and
from XYZ space to a specified RGB colorspace.

The RGB source- or targetcolorspace are defined by their color
gamut, specificially the CIE-1931 chromaticity coordinates of
their red, green and blue primaries and white point.

These helper function are not yet used, but they are the equivalent
of Screen()'s built-in C implementations, and can be used for converting
from one RGB color space to another, e.g., for future HDR applications,
if a different colorspace is about to be used for output than our
current BT-2020 color space for HDR-10.