Screen: Allow override of used eotf for pixelFormat 11 playback.
In pixelFormat 11, now the GStreamer detected EOTF type of the movie
can be overriden with a new optional 'movieOptions' parameter in
Screen('OpenMovie'). movieOptions = 'OverrideEOTF=14' would request
overriding the detected EOTF type to type 14 (PQ).

This allows to play back HDR content with GStreamer versions < 1.18,
which can not detect HDR transfer functions, so we can help out a
bit here. Tested with GStreamer 1.16.2. Ofc. HDR static metadata is
also not returned by older GStreamer, so that info is missing and
may lead to less optimal quality on HDR displays that actually know
what to do with that info. Mine doesn't, so effectively stuff looks
as good with GStreamer 1.16 + this option as on GStreamer 1.18.

PlayMoviesDemo(movie, 2); demonstrates this hack now, by selecting
a type 14 PQ HDR EOTF if hdr flag is selected to be 2. Also some
white-space cleanup in PlayMoviesDemo when we are at it.