Screen: Release external memory handles, as appropriate.
External memory handles, e.g., for OpenGL-Vulkan interop, must be
released in some occassions, e.g., if a new handle is assigned in
PsychSetPipelineExportTextureInteropMemory(), or if an error happens
before the handle can be consumed by the OpenGL implementations
glImportXXX() function.

On MS-Windows we also need to keep track of handles and release them
manually at onscreen window close time in PsychShutdownImagingPipeline(),
but not on Linux/Unix.

For some reason the glImportMemoryFdEXT() function for Unix systems
does take ownership of imported fd handles and takes care of their
life-cycle, but the equivalent MS-Windows glImportMemoryWin32HandleEXT()
function does not take ownership, instead leaving the job to us to
CloseHandle() the handle at the end of a session.

-> This should prevent leaking shared GPU memory between OpenGL and
   Vulkan in case of error, or always on MS-Windows.

-> There's still a small chance of leaks if an error abort would
   hapen in the PsychVulkan.m script between getting the handle
   from PsychVulkanCore() and injecting it into Screen(), but this
   is unlikely if i did my job right, and in practice the leaks
   would be of about ~28 MB per run for a 2560x1440 window or
   ~64 MB for a 4k display. Not negligible, but would still take
   a while to affect performance on todays  multi-GB RAM systems.