PsychVulkan: Need to call 'WindowBackendOverrides' early for 10bpc+
It is too late for reconfiguration of display outputs to >= 10 bpc
once the vulkan fullscreen window is open, as then RandR requests
will be locked out until the Vulkan window is closed again after
end of session.

Do an early call if >= 10 bpc need is anticipated. Do always the
late call, as at that point we actually know the true framebuffer
bit depth, so can set accurate settings.