Screen/Linux: Improve automatically set 'max bpc' for RandR outputs.
Only auto-increase 'max bpc' to 10 bpc iff it is <= 8 bpc, or the
special default value 16 for AMD gpu's. Iow. if the user or 3rd party
tools haven't changed it.

This allows defaults to be upgraded for high precision output, but
leaves the user in control if they want a specific setting.

The reason for this specific strategy can be found in the code comments.

Btw. we don't auto-upgrade to 12 bpc for 11/12 bpc framebuffers,
because true 12 bpc displays are very rare, but the most common
type of 12 bpc displays are HDMI deep color monitors which are
mandated to advertise 12 bpc even if they can only do 10 bpc,
iow. most 12 bpc displays are fake 12 bpc! This is unfortunately
mandated by the HDMI spec!